interceptor-main/storetracker-interface/requirements.txt
paho_mqtt==2.1.0
pyserial==3.5
PyYAML==6.0.2